validacion larga

+++ ProyectoFinal_BD/src/co/edu/unbosque/controller/Controller.java
@@ -745,7 +745,7 @@ public class Controller implements ActionListener {
 	}
 	private static boolean isNumeric(String cadena){
 		try {
-			Integer.parseInt(cadena);
+			Long.parseLong(cadena);
 			return true;
 		} catch (NumberFormatException nfe){
 			return false;
